Convert 'int32_t' to 'int8_t' Space is precious! We have only 64B of RAM.

+++ add.c
@@ -1,7 +1,7 @@
 #include <stdint.h>
 #include "add.h"
 
-int32_t add(int32_t a,int32_t b)
+int8_t add(int8_t a,int8_t b)
 {
    int32_t x=0,y=0,z=0;
    x = a;
